+++ contracts/nft/Cargo.toml
@@ -11,12 +11,12 @@ soroban-sdk = { version = "21.7.7", features = ["testutils"] }
 
 [lib]
 name = "soroban_nft"
-path = "src/soroban_nft.rs"
+path = "src/lib.rs"
 doctest = false
 crate-type = ["lib", "cdylib"]
 
-[build]
-target = "wasm32-unknown-unknown"
+#[build]
+#target = "wasm32-unknown-unknown"
 
 [target.wasm32-unknown-unknown]
 rustflags = ["-C", "target-feature=-reference-types"]
