+++ contracts/nft/packages/soroban_nft/README.md
@@ -8,7 +8,7 @@ This library was automatically generated by Soroban CLI using a command similar 
 soroban contract bindings ts \
   --rpc-url https://soroban-testnet.stellar.org \
   --network-passphrase "Test SDF Network ; September 2015" \
-  --contract-id CAEJTO4RBXU3U336KW4XKNYKNZ72CKIIVXQPNL5KKL7Y6BBSTLY6ACKV \
+  --contract-id CDH2OICCUXRWEY52YRRQDJBUJCKSVEN4JFUKWTLWWXXO5QLZ6Q6ZIGDQ \
   --output-dir ./path/to/soroban_nft
 ```
 
@@ -30,7 +30,7 @@ However, we've actually encountered [frustration](https://github.com/stellar/sor
 
 ```json
 "scripts": {
-  "postinstall": "soroban contract bindings ts --rpc-url https://soroban-testnet.stellar.org --network-passphrase \"Test SDF Network ; September 2015\" --id CAEJTO4RBXU3U336KW4XKNYKNZ72CKIIVXQPNL5KKL7Y6BBSTLY6ACKV --name soroban_nft"
+  "postinstall": "soroban contract bindings ts --rpc-url https://soroban-testnet.stellar.org --network-passphrase \"Test SDF Network ; September 2015\" --id CDH2OICCUXRWEY52YRRQDJBUJCKSVEN4JFUKWTLWWXXO5QLZ6Q6ZIGDQ --name soroban_nft"
 }
 ```
