+++ src/pages/Index.tsx
@@ -131,7 +131,6 @@ const Index = () => {
                   // VIEJO NO SIRVE
                   // Mint the transaction
                   sorobanClient.options.publicKey = publicKey;
-
                   const response = await sorobanClient.mint({ to: publicKey });
 
                   // Set the public key for signing
